requirements.txt
pyyaml
amqtt
sqlalchemy>=1.4.0
aiomyql
uvloop
passlib
